server/requirements.txt
fastapi>=0.114,<1
uvicorn[standard]>=0.30,<1
sqlalchemy>=2,<3
alembic>=1,<2
psycopg[binary]>=3,<4
pydantic>=2,<3
pydantic-settings>=2,<3
python-multipart>=0.0.9,<1
redis>=5,<6
requests>=2,<3
docling>=2,<3
Jinja2>=3,<4
langchain>=0.3,<1
langgraph>=0.2,<1
groq>=0.11,<1
orjson>=3,<4
python-slugify>=8,<9
beautifulsoup4>=4,<5
python-docx>=1,<2
pypdf>=4,<5
Pillow>=10,<11
pytesseract>=0.3.10,<0.4
openpyxl>=3.1,<4
